suppression de mathalea_amc.js
